moved find test
