+++ nucliadb/nucliadb/search/tests/integration/requesters/test_utils.py
@@ -28,7 +28,6 @@ from nucliadb_models.search import SearchOptions, SortField, SortOptions, SortOr
 
 
 @pytest.mark.asyncio
-@pytest.mark.xfail  # pulling start/end position for vectors results needs to be fixed
 async def test_vector_result_metadata(
     search_api: Callable[..., AsyncClient], multiple_search_resource: str
 ) -> None:
